+++ FretFrameworkV2/Serialization/BCHFileWriter.h
@@ -17,10 +17,10 @@ public:
 	virtual void writeDifficultyTrack(unsigned char index) override;
 	virtual void writeAnimationTrack() override;
 	virtual void finishTrack() override;
-	virtual void startEvent(uint32_t position, ChartEvent ev) override;
+	virtual void startEvent(uint64_t position, ChartEvent ev) override;
 	virtual void finishEvent() override;
-	virtual void writeSingleNote(const std::pair<size_t, uint32_t>& note) override;
-	virtual void writeMultiNote(const std::vector<std::pair<size_t, uint32_t>>& notes) override;
+	virtual void writeSingleNote(const std::pair<size_t, uint64_t>& note) override;
+	virtual void writeMultiNote(const std::vector<std::pair<size_t, uint64_t>>& notes) override;
 	virtual void writeSingleNoteMods(const std::vector<char>& mods) override;
 	virtual void writeMultiNoteMods(const std::vector<std::pair<char, size_t>>& mods) override;
 	virtual void writeText(std::string_view str) override;
@@ -28,7 +28,7 @@ public:
 	virtual void writePitch(Pitch<-1, 9> pitch) override;
 
 	virtual void writeLyric(std::pair<size_t, std::string_view> lyric) override;
-	virtual void writePitchAndDuration(const std::pair<Pitch<-1, 9>, uint32_t>& note) override;
+	virtual void writePitchAndDuration(const std::pair<Pitch<-1, 9>, uint64_t>& note) override;
 
 	virtual void writeNoteName(NoteName note, PitchType type) override;
 	virtual void writeLeftHand(size_t position) override;
@@ -41,7 +41,7 @@ private:
 	using BinaryFileWriter::writeTag;
 
 	std::vector<std::streampos> m_trackPositions;
-	uint32_t m_position;
+	uint64_t m_position;
 	void writeTrackHeader(const char(&tag)[5]);
 
 	void writeTrackLength();
